+++ priv/repo/migrations/20200211130308_add_respondents_canonical_phone_number.exs
@@ -1,0 +1,9 @@
+defmodule Ask.Repo.Migrations.AddRespondentsCanonicalPhoneNumber do
+  use Ecto.Migration
+
+  def change do
+    alter table(:respondents) do
+      add :canonical_phone_number, :string
+    end
+  end
+end